feat(jvm): 添加 cglib.jar 及 jvm 练习代码

+++ practice/src/main/java/run_time_data_area/pc_register/ProgramCounterRegisterSimpleDemo.java
@@ -9,7 +9,7 @@ public class ProgramCounterRegisterSimpleDemo {
 
     public static void main(String[] args) {
         int i = 10;
-        int j = 20;
-        int k = i + j;
+        long j = 20;
+        double k = i + j;
     }
 }
